Change Collection to MutableList, Change dni to have same type as the others ids

+++ src/main/kotlin/UNQ/TTIP/GOAT/model/Player.kt
@@ -7,16 +7,16 @@ import javax.persistence.*
 
 @Entity
 @Table(name = "player")
-public class Player (@Id
+public class Player(@Id
                      @Column(name = "dni", nullable = false)
-                     var dni: Int,
-                     var name: String,
-                     var surname: String,
-                     var birth: Date,
+                    var dni: Long,
+                    var name: String,
+                    var surname: String,
+                    var birth: Date,
 
-                     @OneToMany( cascade = [CascadeType.ALL], fetch = FetchType.LAZY)
+                    @OneToMany( cascade = [CascadeType.ALL], fetch = FetchType.LAZY)
                      @JoinColumn(name="player_dni")
-                     var teams : List<PlayerTeamStats> = mutableListOf())
+                    var teams: List<PlayerTeamStats> = mutableListOf())
     :StatsSheet() {
 
     @OneToMany( cascade = [CascadeType.ALL], fetch = FetchType.LAZY)
